Add signup & save user config to file
- Added signup function
before executing `signupFlowVolume()`, make sure that game variable 13 is initialized with the chip ID. The function would automatically assign a random User ID to you and save it in game variable 14.

- Added save config function to fv_config.json
- This would store user_id & chip_id game variables even after game reset/deleted saves
in order to load the config file, you need to `loadFVConfig()` and put a Text/Wait command infront in order to give time to fetch the config.
`saveFVConfig()` is used only if you want to overwrite the values in the fv_config.json file using the game variables.

- Save file turned off BGM